lists

+++ 0x1E-search_algorithms/0-linear.c
@@ -1,4 +1,5 @@
 #include "search_algos.h"
+
 /**
  * linear_search - function that searches for a value in an array of integers
  *					using the Linear search algorithm
@@ -10,18 +11,18 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
+	int i;
 
 	if (!array || !size)
 		return (-1);
 
-	for (i = 0; i < size; i++)
+	for (i = 0; i < (int)size; i++)
 	{
-		printf("Value checked array[%lu] = [%i]\n", i, array[i]);
+		printf("Value checked array[%i] = [%i]\n", i, array[i]);
 		if (array[i] == value)
 			break;
 	}
-	if (i >= size)
+	if (i >= (int)size)
 		return (-1);
 	return (i);
 }
